Update posts tests to be split between api tests and web tests
